Add lots of error checking, but finally found the bug, document ids assignment was wrong

+++ src/preprocess/preprocess.hpp
@@ -20,7 +20,9 @@ struct FastTrie {
     FastTrie();
 
     static inline std::vector<int> get_def();
+
     static inline int new_node();
+
     static void insert(std::string &str, bool val = true);
 
     void start(char c);
@@ -47,13 +49,13 @@ struct Preprocessor {
 
     static inline constexpr bool validChar(char c);
 
-    int processText(data_type &all_data, int docid, int zone, const std::string &text, int start, int end);
+    int processText(data_type &all_data, int docid, int zone, const std::string &text, size_t start, size_t end);
 
     inline static bool fast_equals(const std::string &src, const std::string &target, int pos);
 
     inline static bool fast_equals(const std::string &src, const std::vector<std::string> &targets, int pos);
 
-    std::vector<std::string> getStemmedTokens(const std::string &text, int start, int end);
+    std::vector<std::string> getStemmedTokens(const std::string &text, size_t start, size_t end);
 
     inline std::string stemming(const sb_symbol *word, int len) const;
 };
